FU-81 refactor: group product form components
