+++ src/components/product/product-form.tsx
@@ -4,11 +4,11 @@ import { useState } from "react";
 import { FormProvider, useForm } from "react-hook-form";
 import { Image, Product } from "product-types";
 import * as style from "./product.css";
-import ItemFieldArray from "./item-field-array";
+import ItemFieldArray from "./form/item-field-array";
 import { SubmitButton } from "../buttons/common-buttons";
-import OptionFieldArray from "./option-field-array";
-import ImagesInput from "./image-input";
-import DiscountFieldArray from "./discount-field-array";
+import OptionFieldArray from "./form/option-field-array";
+import ImagesInput from "./form/image-input";
+import DiscountFieldArray from "./form/discount-field-array";
 
 const ProductForm = () => {
   const defaultValues: Product = {
